+++ ruby/lib/cucumber/messages/envelope.rb
@@ -48,6 +48,10 @@ module Cucumber
 
       attr_reader :test_step_started
 
+      attr_reader :test_run_hook_started
+
+      attr_reader :test_run_hook_finished
+
       attr_reader :undefined_parameter_type
 
       def initialize(
@@ -67,6 +71,8 @@ module Cucumber
         test_run_started: nil,
         test_step_finished: nil,
         test_step_started: nil,
+        test_run_hook_started: nil,
+        test_run_hook_finished: nil,
         undefined_parameter_type: nil
       )
         @attachment = attachment
@@ -85,6 +91,8 @@ module Cucumber
         @test_run_started = test_run_started
         @test_step_finished = test_step_finished
         @test_step_started = test_step_started
+        @test_run_hook_started = test_run_hook_started
+        @test_run_hook_finished = test_run_hook_finished
         @undefined_parameter_type = undefined_parameter_type
         super()
       end
@@ -116,6 +124,8 @@ module Cucumber
           test_run_started: TestRunStarted.from_h(hash[:testRunStarted]),
           test_step_finished: TestStepFinished.from_h(hash[:testStepFinished]),
           test_step_started: TestStepStarted.from_h(hash[:testStepStarted]),
+          test_run_hook_started: TestRunHookStarted.from_h(hash[:testRunHookStarted]),
+          test_run_hook_finished: TestRunHookFinished.from_h(hash[:testRunHookFinished]),
           undefined_parameter_type: UndefinedParameterType.from_h(hash[:undefinedParameterType])
         )
       end